feat: 完成核心业务链路闭环（酒店搜索、优惠券、订单及退房退款逻辑）
1. [新增] 酒店搜索与房型匹配：支持按城市、入住/退房时间段动态查询可用库存。
2. [新增] 营销优惠系统：集成优惠券选择逻辑，支持订单金额自动抵扣与计算。
3. [新增] 订单全生命周期管理：模拟支付流转、前台办理入住及订单状态追踪。
4. [优化] 财务结算算法：实现“提前退房”自动差价计算，支持基于天数的阶梯退款逻辑。
5. [重构] 三端角色权限微调：优化普通用户、管理员与酒店管理员的业务操作边界。
6. [同步] 通用图片系统集成：将 target_id/type 架构全面应用至酒店及房型展示。

- 解决了酒店高频查询下的数据一致性问题。
- 确保了退款金额计算的精确度，避免了 Read-Modify-Write 导致的潜在坏账。

+++ src/main/java/com/travelfamilies/controller/CommentController.java
@@ -20,7 +20,7 @@ public class CommentController {
     @PostMapping("/add")
     public Result<?> addComment(@RequestBody AddCommentRequest addCommentRequest, HttpServletRequest httpServletRequest) {
 
-        int userId = (int) httpServletRequest.getAttribute("userID");
+        long userId = (long) httpServletRequest.getAttribute("userID");
         return commentService.addComment(userId, addCommentRequest);
     }
